requirements.txt
django
djangorestframework
django-cors-headers
whitenoise
gunicorn
django-filter
drf-yasg
djangorestframework-simplejwt